backend/requirements.txt
# Web framework
fastapi==0.135.1
uvicorn==0.42.0
starlette==1.0.0
pydantic==2.12.5
python-multipart

# Data
numpy
pandas==3.0.1
scikit-learn==1.8.0
Pillow==12.1.1
